Restructure modules and remove leftover code
