+++ src/database/mod.rs
@@ -3,7 +3,7 @@ use entity::{build, pull_request};
 
 mod sea_orm_client;
 
-use crate::github::api::client::PullRequestNumber;
+use crate::github::PullRequestNumber;
 use crate::github::{CommitSha, GithubRepoName};
 pub use sea_orm_client::SeaORMClient;
