+++ src/lib.rs
@@ -1,8 +1,7 @@
-mod command;
+mod bors;
 mod config;
 pub mod database;
 pub mod github;
-mod handlers;
 mod permissions;
 
 mod tests;
